[enum_name_map_arg_bloc] update `SautModalBottomSheetRouteBuilder`
related commits:
a6d62ca8de2e992f6a4e236198b00a2550a1c451
Support keeping a bottom sheet with a DraggableScrollableSheet from closing on drag/fling to min extent (#127339)

83f19be2502a51e2c1c84584b0eeab53bee3b202
Fix bottom sheet rebuilding when tapping (#127526)

This fixes an issue where the bottom sheet would rebuild when enableDrag is set to true on every tap. This is because DragGestureRecognizer would win the arena by default and dispatch the drag callbacks (in acceptGesture) even though it had not met the drag threshold. This changes keep the default behavior of DragGestureRecognizer the same, but adds a parameter onlyAcceptDragOnThreshold that a user can use to stop drag callbacks from being fired when the drag threshold has not been met.